chore: rename v2
