+++ app/routes/bookmarks/index.tsx
@@ -15,7 +15,7 @@ import type {
   VideoTable,
 } from "../../db/models";
 import { $R, ROUTE_DEF } from "../../misc/routes";
-import { rpcClientQuery } from "../../trpc/client-v2";
+import { rpcClientQuery } from "../../trpc/client";
 import {
   disableUrlQueryRevalidation,
   useTypedUrlQuery,
