+++ app/routes/decks/import.tsx
@@ -4,7 +4,7 @@ import { useMutation } from "@tanstack/react-query";
 import { useForm } from "react-hook-form";
 import { toast } from "react-hot-toast";
 import { R } from "../../misc/routes";
-import { rpcClient } from "../../trpc/client-v2";
+import { rpcClient } from "../../trpc/client";
 import { cls } from "../../utils/misc";
 import type { PageHandle } from "../../utils/page-handle";
