+++ app/routes/users/signin.tsx
@@ -2,7 +2,7 @@ import { Link } from "@remix-run/react";
 import { useMutation } from "@tanstack/react-query";
 import { useForm } from "react-hook-form";
 import { $R } from "../../misc/routes";
-import { rpcClientQuery } from "../../trpc/client-v2";
+import { rpcClientQuery } from "../../trpc/client";
 import { encodeFlashMessage } from "../../utils/flash-message";
 import { cls } from "../../utils/misc";
 import { navigateRefresh } from "../../utils/misc-client";
