+++ app/routes/videos/$id.tsx
@@ -19,7 +19,7 @@ import { PopoverSimple } from "../../components/popover";
 import { E, T, TT, selectOne } from "../../db/drizzle-client.server";
 import type { UserTable, VideoTable } from "../../db/models";
 import { $R, ROUTE_DEF } from "../../misc/routes";
-import { rpcClientQuery } from "../../trpc/client-v2";
+import { rpcClientQuery } from "../../trpc/client";
 import { useDocumentEvent } from "../../utils/hooks-client-utils";
 import { intl } from "../../utils/intl";
 import {
